feature: prevent double addtion/deletion of friend

+++ server/controllers/friendsController.js
@@ -14,24 +14,35 @@ exports.getAll = async (req, res) => {
 exports.addFriend = async (req, res) => {
   const userId = req.user.userId;
   const { friendId } = req.body;
-  try {
-    await Friend.addFriend(userId, friendId);
-    return res.json({ success: true, message: "Succesfully added friend" });
-  } catch {
-    console.log(err);
-    return res.status(500).json({ success: false, message: "Something went wrong" });
+  const alreadyExists = await Friend.checkIfExists(userId, friendId);
+
+  if (!alreadyExists) {
+    try {
+      await Friend.addFriend(userId, friendId);
+      return res.json({ success: true, message: "Succesfully added friend" });
+    } catch {
+      console.log(err);
+      return res.status(500).json({ success: false, message: "Something went wrong" });
+    }
+  } else {
+    return res.status(400).json({ success: false, message: "Friend already exists" });
   }
 };
 
 exports.deleteFriend = async (req, res) => {
   const userId = req.user.userId;
   const friendId = req.params.friendId;
+  const alreadyExists = await Friend.checkIfExists(userId, friendId);
 
-  try {
-    await Friend.deleteById(userId, friendId);
-    return res.json({ success: true, message: "Successfully removed friend" });
-  } catch (err) {
-    console.log(err);
-    return res.status(500).json({ success: false, message: "Something went wrong" });
+  if (alreadyExists) {
+    try {
+      await Friend.deleteById(userId, friendId);
+      return res.json({ success: true, message: "Successfully removed friend" });
+    } catch (err) {
+      console.log(err);
+      return res.status(500).json({ success: false, message: "Something went wrong" });
+    }
+  } else {
+    return res.status(400).json({ success: false, message: "Friend does not exist" });
   }
 };
